Added verify_tomato_potato_homology.
Another protein_output_dir_path was added in the main function to be used in the call of verify_tomato_potato_homology. The gene_seq_to_protein function still computes this path on its own and creates the directory but if this function fails the verify function call might fail because the protein sequences are not created. It should be checked.